// tools/toml-to-json-converter/client/types.ts
export type TomlToJsonConverterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  importFromFileLabel: string
  downloadJsonLabel: string
  tomlLabel: string
  tomlDescription: string
  jsonLabel: string
  jsonDescription: string
  jsonEmptyDescription: string
  tomlPlaceholder: string
  invalidTomlLabel: string
  copyJsonLabel: string
  copiedLabel: string
}>